Even more cleanup & fixes
